core: added primitive literal suffixes

+++ qc-core/include/qc-core/arena.hpp
@@ -156,9 +156,9 @@ namespace qc
             u32 refN;
         };
 
-        static constexpr u64 _maxPageN{u64(1u) << 16};
+        static constexpr u64 _maxPageN{1u << 16};
         static constexpr u64 _maxCapacity{_maxPageN * pageSize - sizeof(_Chunk)};
-        static constexpr u64 _maxValChunkN{(u64(1u) << 16) - 1u};
+        static constexpr u64 _maxValChunkN{(1u << 16) - 1u};
 
         static u64 _pageN(u64 capacity);
